- Refactored code, moved yaml classes at yaml package level - Added support for specifying schema for input if provided
