+++ sope-etl/src/test/scala/com/sope/etl/YamlTransformWithoutSourceTest.scala
@@ -4,7 +4,7 @@ import java.sql.{Date => SDate}
 
 import com.sope.etl.TestContext._
 import com.sope.etl.model.{Date, Product, ProductDim, Transactions}
-import com.sope.etl.yaml.YamlFile.IntermediateYaml
+import com.sope.etl.yaml.IntermediateYaml
 import org.scalatest.{FlatSpec, Matchers}
 
 /**
